fin seccion 13

+++ src/app/api/todos/route.ts
@@ -1,5 +1,6 @@
+import { getUserSessionServer } from "@/auth/actions/auth-actions";
 import prisma from "@/lib/prisma";
-import { NextResponse } from "next/server";
+import { NextResponse, NextRequest } from "next/server";
 import * as yup from "yup";
 
 export async function GET(request: Request) {
@@ -8,14 +9,22 @@ export async function GET(request: Request) {
    const skip = Number(searchParams.get("skip") ?? "0");
 
    if (isNaN(take)) {
-      return NextResponse.json({ error: "Invalid take" }, { status: 400 });
+      return NextResponse.json(
+         { message: "Take tiene que ser un número" },
+         { status: 400 }
+      );
    }
+
    if (isNaN(skip)) {
-      return NextResponse.json({ error: "Invalid skip" }, { status: 400 });
+      return NextResponse.json(
+         { message: "Skip tiene que ser un número" },
+         { status: 400 }
+      );
    }
+
    const todos = await prisma.todo.findMany({
-      take,
-      skip,
+      take: take,
+      skip: skip,
    });
 
    return NextResponse.json(todos);
@@ -26,28 +35,40 @@ const postSchema = yup.object({
    complete: yup.boolean().optional().default(false),
 });
 
-export async function POST(req: Request, res: Response) {
+export async function POST(request: Request) {
+   const user = await getUserSessionServer();
+
+   if (!user) {
+      return NextResponse.json("No autorizado", { status: 401 });
+   }
+
    try {
-      const body = await postSchema.validate(await req.json());
+      const { complete, description } = await postSchema.validate(
+         await request.json()
+      );
 
       const todo = await prisma.todo.create({
-         data: body,
+         data: { complete, description, userId: user.id },
       });
+
       return NextResponse.json(todo);
    } catch (error) {
       return NextResponse.json(error, { status: 400 });
    }
 }
+export async function DELETE(request: Request) {
+   const user = await getUserSessionServer();
 
-export async function DELETE(req: Request, res: Response) {
+   if (!user) {
+      return NextResponse.json("No autorizado", { status: 401 });
+   }
+
    try {
       await prisma.todo.deleteMany({
-         where: {
-            complete: true,
-         },
+         where: { complete: true, userId: user.id },
       });
       return NextResponse.json("Borrados");
    } catch (error) {
-      return NextResponse.json({ error }, { status: 400 });
+      return NextResponse.json(error, { status: 400 });
    }
 }
